backend/requirements.txt
fastapi==0.115.6
python-multipart>=0.0.9
uvicorn[standard]==0.34.0
httpx==0.28.1
pydantic==2.10.4
python-dotenv==1.0.1
sse-starlette==2.2.1
qiskit<2.0.0
qiskit-aer>=0.17.0
pylatexenc==2.10
matplotlib>=3.8.0
neo4j
falkordb
graphiti
networkx
dimod
torch>=2.1.0
torchvision
torchaudio
qiskit-algorithms
qiskit-machine-learning
qiskit-ibm-runtime
# Cloud AI providers
anthropic>=0.40.0
openai>=1.0.0
google-genai>=1.0.0
# Native Apple Silicon (Host execution only)
mlx-lm; sys_platform == 'darwin'
google-genai>=1.0.0
# Vector search
chromadb>=0.5.0
sentence-transformers>=3.0.0
# D-Wave simulated annealing (local solver)
dwave-neal>=0.6.0
# Database & Cache (Batch 9 & 10)
sqlalchemy>=2.0.0
alembic>=1.13.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
celery>=5.4.0
redis>=5.0.0

# Enterprise Auth
python-keycloak>=3.8.0
python-jose[cryptography]==3.3.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0

yahooquery>=2.3.0
websockets>=12.0
slowapi>=0.1.9
